Moved the calendar functionality to its own component so it can be used in the new date range component.
